feat(cmds): add load and apps commands

+++ scripts/wallpaper_core.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+# TODO: Add a option to cache reset so it deleted the wallpapers saved in cache and
+# the cache of the wallpaper frames to generate new ones
+#
+# also add a command to add a wallpaper inside the wallpaper menu
+
 source "$RETRO_DIR/scripts/lib/battery.sh"
 source "$RETRO_DIR/scripts/lib/variable.sh"
 
@@ -32,58 +37,52 @@ set_wallpaper() {
     if [[ -f $input_path ]]; then
         wall_path="$input_path"
     elif [[ -f "$WALL_DIR/$input_path" ]]; then
-        wall_path="$WALL_DIR/$input_path"
+        wall_path="$wall_path"
     else
         return 1
     fi
 
     local filename=$(basename "$wall_path")
-    local engine=$(get_var "CLR_ENGINE")
-    local is_saver=$(get_var "BAT_SAVER_ACTIVE")
-    local force_static=$(get_var "WALL_STATIC_FORCED")
-    local static_on_bat=$(get_var "WALL_STATIC_ON_BAT")
-    local on_bat=$(is_on_battery)
-
-    : ${engine:="matugen"}
-    : ${is_saver:="false"}
-    : ${force_static:="false"}
-    : ${static_on_bat:="false"}
-
     local is_video=false
     [[ $filename =~ \.(mp4|mkv|webm)$ ]] && is_video=true
 
-    local should_be_static=false
-    if [[ $is_saver == "true" ]]; then
-        should_be_static=true
-    elif [[ $force_static == "true" ]]; then
-        should_be_static=true
-    elif [[ $on_bat == "true" && $static_on_bat == "true" ]]; then
-        should_be_static=true
+    local is_first_load=false
+    if ! pgrep -x "awww-daemon" >/dev/null; then
+        is_first_load=true
+        nohup awww-daemon >/dev/null 2>&1 &
+        sleep 0.1
     fi
 
-    local color_source=$(generate_cache "$wall_path")
-
-    if [[ $engine == "matugen" ]]; then
-        matugen image -b wal "$color_source" --source-color-index 0
-    else
-        wal -i "$color_source" -n -q -e
-    fi
-
-    pkill mpvpaper
-
+    local static_source="$wall_path"
     if [[ $is_video == "true" ]]; then
-        if [[ $should_be_static == "true" ]]; then
-            awww img "$color_source" --transition-type fade --transition-duration 0
-        else
-            pgrep -x "awww-daemon" >/dev/null || awww-daemon &
-            awww img "$color_source" --transition-type grow --transition-duration 2.5 --transition-fps 120
-
-            sleep 2.2
-            NV_PRIME_RENDER_OFFLOAD=0 mpvpaper -o "--loop --hwdec=vaapi --panscan=1.0 --no-audio" "*" "$wall_path" &
-        fi
-    else
-        awww img "$wall_path" --transition-type grow --transition-duration 2.5 --transition-fps 120
+        static_source=$(generate_cache "$wall_path")
     fi
+
+    if [[ $is_first_load == "true" ]]; then
+        awww img "$static_source" --transition-type fade --transition-duration 0
+    else
+        awww img "$static_source" --transition-type grow --transition-duration 2.5 --transition-fps 120
+    fi
+
+    (
+        matugen image -b wal "$static_source" --source-color-index 0 >/dev/null 2>&1
+
+        if [[ $is_video == "true" ]]; then
+            local is_saver=$(get_var "BAT_SAVER_ACTIVE" "false")
+            local force_static=$(get_var "WALL_STATIC_FORCED" "false")
+            local static_on_bat=$(get_var "WALL_STATIC_ON_BAT" "false")
+            local on_bat=$(is_on_battery)
+
+            if [[ $is_saver == "false" && $force_static == "false" ]]; then
+                if [[ $on_bat == "false" || $static_on_bat == "false" ]]; then
+                    pkill mpvpaper
+                    [[ $is_first_load == "false" ]] && sleep 2.2
+
+                    NV_PRIME_RENDER_OFFLOAD=0 mpvpaper -o "--loop --hwdec=vaapi --panscan=1.0 --no-audio" "*" "$wall_path" &
+                fi
+            fi
+        fi
+    ) &
 
     set_var "WALL_CURRENT" "$wall_path"
 }
